server: delete 'gortex server'; folded into 'gortex daemon --http'
The standalone server command is removed — its /v1 REST surface now rides
on the daemon's HTTP listener (prior commit), so a single daemon process
serves both transports and there is no second graph-owning process to
keep in sync. The two still-shared helpers (isLocalhostBind,
resolveServerID) move to http_helpers.go; the agent-facing HTTP-transport
doc is updated to describe the unified daemon surface. This leaves
serverstack.NewSharedServer as the single graph/indexer/engine/MCP wiring
site.
