+++ internal/agents/instructions.go
@@ -450,7 +450,7 @@ HTTP transport mirrors the surface at ` + "`/v1/overlay/sessions/*`" + `; the ` 
 
 ### MCP 2026 Streamable HTTP transport (` + "`/mcp`" + `)
 
-` + "`gortex server`" + ` and ` + "`gortex daemon --http-addr <addr>`" + ` both expose the **MCP 2026 Streamable HTTP transport** on ` + "`POST /mcp`" + `, ` + "`GET /mcp`" + ` (SSE upstream for server-initiated notifications), ` + "`DELETE /mcp`" + ` (session termination), ` + "`OPTIONS /mcp`" + ` (CORS preflight). One endpoint, one wire format — the spec the June 2026 MCP release locks in. Stateless per request: every POST carries ` + "`Mcp-Session-Id`" + ` and the transport replays state out of a ` + "`streamable.SessionStore`" + ` (in-memory by default, TTL-evicted; swap in a shared backend like Redis to run multiple workers behind a load balancer). ` + "`initialize`" + ` mints the id and returns it on the response header; an unknown id replies with a JSON-RPC ` + "`-32001 session not found`" + ` envelope. JSON-RPC batches preserve order; notification-only batches return HTTP 202. ` + "`tools/call`" + ` frames flow through the same multi-server router that serves ` + "`/v1/tools/<name>`" + `, so workspace scoping carries over unchanged. ` + "`gortex daemon`" + ` enables it via ` + "`--http-addr 127.0.0.1:7411 [--http-auth-token <token>]`" + `; non-localhost binds require an auth token (or ` + "`$GORTEX_DAEMON_HTTP_TOKEN`" + `). ` + "`/healthz`" + ` is exempt from auth so liveness probes work. ` + "`gortex server`" + ` mounts ` + "`/mcp`" + ` unconditionally alongside the legacy ` + "`/v1/*`" + ` surface.
+` + "`gortex daemon --http-addr <addr>`" + ` exposes the **MCP 2026 Streamable HTTP transport** on` + "`POST /mcp`" + `, ` + "`GET /mcp`" + ` (SSE upstream for server-initiated notifications), ` + "`DELETE /mcp`" + ` (session termination), ` + "`OPTIONS /mcp`" + ` (CORS preflight). One endpoint, one wire format — the spec the June 2026 MCP release locks in. Stateless per request: every POST carries ` + "`Mcp-Session-Id`" + ` and the transport replays state out of a ` + "`streamable.SessionStore`" + ` (in-memory by default, TTL-evicted; swap in a shared backend like Redis to run multiple workers behind a load balancer). ` + "`initialize`" + ` mints the id and returns it on the response header; an unknown id replies with a JSON-RPC ` + "`-32001 session not found`" + ` envelope. JSON-RPC batches preserve order; notification-only batches return HTTP 202. ` + "`tools/call`" + ` frames flow through the same multi-server router that serves ` + "`/v1/tools/<name>`" + `, so workspace scoping carries over unchanged. ` + "`gortex daemon`" + ` enables it via ` + "`--http-addr 127.0.0.1:7411 [--http-auth-token <token>]`" + `; non-localhost binds require an auth token (or ` + "`$GORTEX_DAEMON_HTTP_TOKEN`" + `). ` + "`/healthz`" + ` is exempt from auth so liveness probes work. When ` + "`--http-addr`" + ` is set the daemon serves both ` + "`/mcp`" + ` and the ` + "`/v1/*`" + ` REST surface on one listener.
 
 ### Speculative Execution (Simulation Sessions)
